+++ jason/src/peer/mod.rs
@@ -2,6 +2,7 @@ mod ice_server;
 mod peer_con;
 mod repo;
 
+#[doc(inline)]
 pub use self::{
     peer_con::{
         Id as PeerId, PeerConnection, PeerEvent, PeerEventHandler, Sdp,
